# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlianHub API",
    "version": "v2",
    "description": "Paths, methods, auth and tags only. The full reference is docs/API.md. Generated by npm run api:doc; do not edit by hand.",
    "x-built-from": "836a8cf4392e"
  },
...
  "paths": {
...
    "/api/v2/custom-fields/ai/jobs/{jobId}": {
      "get": {
        "summary": "Any signed-in member of the workspace, and only for a job that person started; any other id answers 404 `Job not found.`. Answers with `data`: the job and its progress.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "read"
      }
    },
    "/api/v2/custom-fields/compute": {
      "post": {
        "summary": "Needs the permission to change a task's custom field. Reads `taskIds` (a list of 1 to 200 ids) and `scope` (`sprint` or left out) from the body; only tasks the caller can open are computed. Answers with `data`: `updated`, the computed `values` and `errors`.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/custom-fields/formula/scope": {
      "get": {
        "summary": "Any signed-in member of the workspace may call it. Reads `projectId` (id) and `taskId` (id, used only when the caller can open the task) from the query. Answers with `data`: `names` a formula may use, with the task's values when a task is named, and the lists of `functions` and `rollupFunctions`.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "read"
      }
    },
    "/api/v2/custom-fields/formula/validate": {
      "post": {
        "summary": "Any signed-in member of the workspace may call it. Reads `expression` (string, required), `fieldId`, `fieldTitle`, `projectId` (ids and string) and `sample` (an object of values to try it on) from the body. A formula that does not parse or that closes a cycle answers a failed status with a `code`. Answers with `data`: `references`, `functions`, `preview` and `previewError`.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/custom-fields/links/resolve": {
      "post": {
        "summary": "Returns, for each task named, the tasks its relationship fields link and the tally of its voting fields.",
        "tags": [
          "Custom fields"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": true,
        "x-token-scope": "write"
      }
    },
    "/api/v2/custom-fields/{fieldId}/ai/apply": {
      "post": {
        "summary": "Any signed-in member of the workspace who can edit that field on the task. Reads `proposalIds` (a list of up to 5 ids from a preview the same person made) from the body; an expired one is refused. Answers with `data`: `applied` (task ids) and `refused` (with a reason each).",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/custom-fields/{fieldId}/ai/jobs": {
      "post": {
        "summary": "Any signed-in member of the workspace who can edit that field on the tasks. Reads `taskIds` (a list of ids, up to the workspace's bulk limit) and `proposalIds` (a list of up to 5 ids) from the body. The fill runs in the background. Answers with `data`: the job.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/custom-fields/{fieldId}/ai/preview": {
      "post": {
        "summary": "Any signed-in member of the workspace who can edit that field on the task; a task they cannot edit it on comes back with a `reason` instead. Reads `taskIds` (a list of 1 to 5 ids) from the body. Nothing is written. Answers with `data.proposals`: for each task, a `proposalId`, the proposed `text` and `fieldValue`.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/custom-fields/{fieldId}/delete": {
      "post": {
        "summary": "Needs the same rights as the usage read; a connected agent is refused. Takes the field and every value of it away for good; archiving (the field update with `isDelete`) hides a field and keeps its values. A field that a rollup or a formula reads answers 409 with the code `FIELD_IS_READ` and the same `readBy` and `readElsewhere`. Answers with `data.tasks`: how many tasks the caller could open held a value.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "write"
      }
    },
    "/api/v2/custom-fields/{fieldId}/usage": {
      "get": {
        "summary": "Needs the permission to edit custom fields in the field's projects, and the field setting for a field of every project. Answers with `data`: `tasks` (how many tasks the caller can open hold a value), `partial` (true when the caller cannot open every task), `readBy` (the titles of the rollups and formulas that read this field) and `readElsewhere` (true when a field of a project the caller cannot open reads it). An unknown id answers 404.",
        "tags": [
          "Not documented in full"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": false,
        "x-token-scope": "read"
      }
    },
    "/api/v2/custom-fields/{fieldId}/vote": {
      "post": {
        "summary": "Casts or withdraws the caller's own vote on a voting field of a task.",
        "tags": [
          "Custom fields"
        ],
        "security": [
          {
            "apiToken": [],
            "companyId": []
          }
        ],
        "parameters": [
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The standard response body."
          }
        },
        "x-stability": "beta",
        "x-documented": true,
        "x-token-scope": "write"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A personal access token."
      },
      "companyId": {
        "type": "apiKey",
        "in": "header",
        "name": "companyid",
        "description": "The workspace the request is for."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "The web app's signed-in session."
      }
    }
  }
}